Reference

ITensorNetworksNext.AbstractNetworkIterator — Type

abstract type AbstractNetworkIterator

A stateful iterator with two states: increment! and compute!. Each iteration begins with a call to increment! before executing compute!, however the initial call to iterate skips the increment! call as it is assumed the iterator is initalized such that this call is implict. Termination of the iterator is controlled by the function done.

ITensorNetworksNext.EachRegion — Type

struct EachRegion{SweepIterator} <: AbstractNetworkIterator

Adapter that flattens each region iterator in the parent sweep iterator into a single iterator.

ITensorNetworksNext.IncrementOnly — Type

struct IncrementOnly{S<:AbstractNetworkIterator} <: AbstractNetworkIterator

Iterator wrapper whos compute! function simply returns itself, doing nothing in the process. This allows one to manually call a custom compute! or insert their own code it in the loop body in place of compute!.

ITensorNetworksNext.RegionIterator — Type

struct RegionIterator{Problem, RegionPlan} <: AbstractNetworkIterator